+++ src/psiutils/widgets.py
@@ -23,16 +23,16 @@ class PsiText(tk.Text):
         tk.Text.__init__(self, *args, **kwargs)
 
         # create a proxy for the underlying widget
-        self._orig = f"{self._w}_orig"
-        self.tk.call("rename", self._w, self._orig)
+        self._orig = f'{self._w}_orig'
+        self.tk.call('rename', self._w, self._orig)
         self.tk.createcommand(self._w, self._proxy)
 
     def _proxy(self, command, *args):
         cmd = (self._orig, command) + args
         result = self.tk.call(cmd)
 
-        if command in ("insert", "delete", "replace"):
-            self.event_generate("<<TextModified>>")
+        if command in ('insert', 'delete', 'replace'):
+            self.event_generate('<<TextModified>>')
 
         return result
 
@@ -349,7 +349,7 @@ class Tooltip:
 
         x, y = tip_pos_calculator(widget, label)
 
-        self.tw.wm_geometry("+%d+%d" % (x, y+self.vertical_offset))
+        self.tw.wm_geometry('+%d+%d' % (x, y+self.vertical_offset))
         # self.tw.wm_geometry(f'{x}x{y}')
 
     def hide(self):