+++ HISTORY.md
@@ -1,5 +1,15 @@
 # Version  history for psiutils
 
+Version 0.2.24 - 05 February 2026
+
+1. Refactor DatePicker and TimePicker
+------------------------------
+
+Version 0.2.23 - 03 February 2026
+
+1. Add DatePicker and TimePicker
+------------------------------
+
 ## 0.2.18 - 22 November 2025
 
 1. Add backup text and replace _ in icon names
@@ -342,5 +352,7 @@
 ## 0.1.0 - 7 October 2024
 
 1. Initial version
+
+